Extract fonts used by example display and font size calibration ..
.. into separate module.
Fonts are registered via AppDescriptor.start() instead of static class
initialization.
This makes the fonts available to all tools, instead of depending on
first opening a display.

In principle also allows building a site-specific product without the
Liberation fonts, but note that this will disable the font size
calibration used when reading *.opi files.

#1093
