+++ app/display/fonts/pom.xml
@@ -1,0 +1,16 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.phoebus</groupId>
+    <artifactId>app-display</artifactId>
+    <version>4.6.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>app-display-fonts</artifactId>
+  <dependencies>
+    <dependency>
+      <groupId>org.phoebus</groupId>
+      <artifactId>core-ui</artifactId>
+      <version>4.6.1-SNAPSHOT</version>
+    </dependency>
+  </dependencies>
+</project>
