+++ app/display/fonts/src/main/resources/META-INF/services/org.phoebus.framework.spi.AppDescriptor
@@ -1,0 +1,1 @@
+org.csstudio.display.fonts.DisplayFontsApp
