+++ app/display/fonts/src/main/resources/fonts/AUTHORS
