+++ app/display/fonts/src/main/resources/fonts/LiberationMono-Bold.ttf
