+++ app/display/fonts/src/main/resources/fonts/LiberationMono-BoldItalic.ttf
