+++ app/display/fonts/src/main/resources/fonts/LiberationMono-Italic.ttf
