+++ app/display/fonts/src/main/resources/fonts/LiberationMono-Regular.ttf
