+++ app/display/fonts/src/main/resources/fonts/LiberationSans-Bold.ttf
