+++ app/display/fonts/src/main/resources/fonts/LiberationSans-BoldItalic.ttf
