+++ app/display/fonts/src/main/resources/fonts/LiberationSans-Italic.ttf
