+++ app/display/fonts/src/main/resources/fonts/LiberationSans-Regular.ttf
