+++ app/display/fonts/src/main/resources/fonts/LiberationSerif-Bold.ttf
