+++ app/display/fonts/src/main/resources/fonts/LiberationSerif-BoldItalic.ttf
