+++ app/display/fonts/src/main/resources/fonts/LiberationSerif-Italic.ttf
