+++ app/display/fonts/src/main/resources/fonts/LiberationSerif-Regular.ttf
