+++ app/display/pom.xml
@@ -11,6 +11,7 @@
     <module>model</module>
     <module>representation</module>
     <module>representation-javafx</module>
+    <module>fonts</module>
     <module>runtime</module>
     <module>editor</module>
     <module>convert-medm</module>
