+++ app/display/representation-javafx/src/main/java/org/csstudio/display/builder/representation/javafx/JFXUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015-2016 Oak Ridge National Laboratory.
+ * Copyright (c) 2015-2020 Oak Ridge National Laboratory.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -36,7 +36,6 @@ public class JFXUtil extends org.phoebus.ui.javafx.JFXUtil
     {
         try
         {
-            CommonFonts.install();
             font_calibration = new JFXFontCalibration().getCalibrationFactor();
         }
         catch (Exception ex)
