+++ core/launcher/src/main/resources/logging.properties
@@ -64,12 +64,14 @@ org.phoebus.applications.pvtree.level = WARNING
 org.phoebus.applications.viewer3d.level = WARNING
 org.phoebus.archive.reader.level = WARNING
 org.csstudio.javafx.rtplot.level = CONFIG
+org.csstudio.apputil.formula.level = WARNING
 org.csstudio.trends.databrowser3.level = WARNING
 org.csstudio.display.builder.level = WARNING
 org.csstudio.display.builder.model.level = WARNING
 org.csstudio.display.builder.representation.level = WARNING
 org.csstudio.display.builder.editor.level = WARNING
 org.csstudio.display.builder.runtime.level = WARNING
+org.csstudio.display.fonts.level = WARNING
 org.csstudio.display.converter.medm.level = INFO
 org.csstudio.display.converter.edm.level = INFO
 org.csstudio.opibuilder.converter.model.level = INFO
