+++ phoebus-product/pom.xml
@@ -74,6 +74,11 @@
     </dependency>
     <dependency>
       <groupId>org.phoebus</groupId>
+      <artifactId>app-display-fonts</artifactId>
+      <version>4.6.1-SNAPSHOT</version>
+    </dependency>
+    <dependency>
+      <groupId>org.phoebus</groupId>
       <artifactId>app-display-runtime</artifactId>
       <version>4.6.1-SNAPSHOT</version>
     </dependency>
